# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CortexCloud API",
    "version": "1.0.0",
    "description": "Production-ready OpenAI-compatible AI gateway with x402 payment-gated routes. Agents pay per call in USDC on Base (eip155:8453) via the x402 protocol \u2014 no API key or subscription required.",
    "contact": {
      "email": "ampwerajonathan50@gmail.com"
    },
...
  },
  "servers": [
    {
      "url": "https://api.cortexcloud.org"
    }
  ],
  "paths": {
...
    "/x402/v1/responses": {
      "post": {
        "operationId": "responses",
        "summary": "OpenAI Responses API completion (x402 payment-gated)",
        "description": "OpenAI Responses API endpoint. Returns HTTP 402 with an x402 PaymentRequirements challenge (USDC on Base) unless a valid X-PAYMENT header is supplied.",
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.005"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "model": {
                    "type": "string"
                  },
                  "input": {
                    "type": "string"
                  },
                  "stream": {
                    "type": "boolean"
                  },
                  "temperature": {
                    "type": "number"
                  },
                  "max_output_tokens": {
                    "type": "integer"
                  }
                },
                "required": [
                  "model",
                  "input"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Result (after valid x402 payment)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "id": "resp_abc",
                    "object": "response",
                    "output": [
                      {
                        "type": "message",
                        "role": "assistant",
                        "content": [
                          {
                            "type": "output_text",
                            "text": "..."
                          }
                        ]
                      }
                    ]
                  }
                }
              }
            }
          },
          "402": {
            "description": "Payment Required \u2014 x402 PaymentRequirements challenge"
          }
        }
      }
    },
...
    "/x402/v1/search": {
      "post": {
        "operationId": "exa_search",
        "summary": "Web search via Exa AI (x402 payment-gated)",
        "description": "Returns HTTP 402 with an x402 PaymentRequirements challenge (USDC on Base) unless a valid X-PAYMENT header is supplied.",
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.03"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "query": {
                    "type": "string",
                    "description": "Search query"
                  },
                  "numResults": {
                    "type": "integer",
                    "description": "Number of results to return",
                    "default": 10
                  },
                  "useAutoprompt": {
                    "type": "boolean",
                    "description": "Use autoprompt for better results"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "neural",
                      "keyword"
                    ],
                    "description": "Search type"
                  },
                  "includeDomains": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Domains to include"
                  },
                  "excludeDomains": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Domains to exclude"
                  },
                  "startPublishedDate": {
                    "type": "string",
                    "description": "Start date for published content (ISO format)"
                  },
                  "endPublishedDate": {
                    "type": "string",
                    "description": "End date for published content (ISO format)"
                  }
                },
                "required": [
                  "query"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Result (after valid x402 payment)"
          },
          "402": {
            "description": "Payment Required \u2014 x402 PaymentRequirements challenge"
          }
        }
      }
    },
    "/x402/v1/search/contents": {
      "post": {
        "operationId": "exa_search_contents",
        "summary": "Fetch content for search result IDs via Exa AI (x402 payment-gated)",
        "description": "Returns HTTP 402 with an x402 PaymentRequirements challenge (USDC on Base) unless a valid X-PAYMENT header is supplied.",
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.02"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Result IDs to fetch content for"
                  },
                  "text": {
                    "type": "boolean",
                    "description": "Include text content",
                    "default": true
                  },
                  "summary": {
                    "type": "boolean",
                    "description": "Include summary",
                    "default": true
                  },
                  "highlights": {
                    "type": "boolean",
                    "description": "Include highlights",
                    "default": false
                  }
                },
                "required": [
                  "ids"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Result (after valid x402 payment)"
          },
          "402": {
            "description": "Payment Required \u2014 x402 PaymentRequirements challenge"
          }
        }
      }
    }
  }
}